refactor(oma-console): move progress bar styles into the oma binary
Move the progress bar style definitions and the progress bar rendering
logic from oma-console into the oma binary itself, and drop the
progress_bar_style feature gate from oma-console.

Also unify the duplicate HumanBytes types: use
oma_utils::human_bytes::HumanBytes everywhere instead of
indicatif::HumanBytes, since they are the same number_prefix-based
implementation.
