+++ src/core/commit_changes.rs
@@ -5,7 +5,6 @@ use bon::Builder;
 use dialoguer::{Confirm, theme::ColorfulTheme};
 use flume::unbounded;
 use jiff::Timestamp;
-use oma_console::indicatif::HumanBytes;
 use oma_history::{DATABASE_PATH, HistoryInfo};
 use oma_logger::{debug, error, info, warn};
 use oma_pager::PagerExit;
@@ -15,6 +14,7 @@ use oma_pm::{
     oma_apt::{self, PackageSort},
     sort::SummarySort,
 };
+use oma_utils::human_bytes::HumanBytes;
 
 use crate::color::Colorize;
 #[cfg(feature = "aosc")]
