+++ src/subcommand/clean.rs
@@ -10,8 +10,8 @@ use crate::fl;
 use crate::success;
 use clap::Args;
 use fs_extra::dir::get_size;
-use oma_console::indicatif::HumanBytes;
 use oma_logger::{debug, error, info};
+use oma_utils::human_bytes::HumanBytes;
 use oma_pm::apt::{OmaApt, OmaAptArgs};
 
 use crate::{error::OutputError, root::root};
