+++ src/subcommand/mirror.rs
@@ -25,14 +25,13 @@ use inquire::ui::Color;
 use inquire::ui::RenderConfig;
 use inquire::ui::StyleSheet;
 use inquire::ui::Styled;
-use oma_console::indicatif::HumanBytes;
-use oma_console::indicatif::ProgressStyle;
+use indicatif::ProgressStyle;
 use oma_logger::{error, info, warn};
 use oma_mirror::MirrorManager;
 use oma_mirror::parser::MirrorConfig;
 use oma_refresh::inrelease::Release;
 use oma_topics::TopicManager;
-use oma_utils::concat_url;
+use oma_utils::{concat_url, human_bytes::HumanBytes};
 use oma_utils::dpkg::dpkg_arch;
 use rayon::iter::IntoParallelRefIterator;
 use rayon::iter::ParallelIterator;
