+++ src/subcommand/size_analyzer/mod.rs
@@ -9,12 +9,11 @@ use std::time::Duration;
 use anyhow::anyhow;
 use clap::Args;
 use dialoguer::console::style;
-use oma_console::indicatif::HumanBytes;
 use oma_pager::{exit_tui, prepare_create_tui};
 use oma_pm::apt::{OmaApt, OmaAptArgs};
 use oma_pm::oma_apt::{Package, PackageSort};
 use oma_pm::pkginfo::OmaPackageWithoutVersion;
-use oma_utils::is_termux;
+use oma_utils::{human_bytes::HumanBytes, is_termux};
 
 use oma_logger::info;
 use std::io::Write;
